database_assistant/requirements.txt
requests
Flask==3.1.0
openai
tabulate==0.9.0
python-dotenv==1.0.0
sqlparse==0.4.4
psycopg2-binary
